refactor(advanced): cart 관련 기능을 Jotai 상태 관리로 마이그레이션
- useCartStore 훅 제거 및 Jotai atoms로 카트 관리 통합
- 새로운 엔티티 구조에 맞게 import 경로 조정
- 카트 관련 계산 로직을 entities/cart로 이동
