+++ src/advanced/entities/cart/index.ts
@@ -1,0 +1,8 @@
+export {
+  cartAtom,
+  addToCartAtom,
+  updateToCartAtom,
+  removeFromCartAtom,
+  resetCartAtom,
+} from './model/atom';
+export { calculateItemTotalQuantity, calculateItemTotal, calculateCartTotal } from './lib';
