+++ src/advanced/hooks/useCouponService.ts
@@ -3,9 +3,9 @@ import { useCallback, useState } from 'react';
 
 import type { CartItem, Coupon } from '../../types';
 import { MINIMUM_ORDER_AMOUNT } from '../constants';
+import { calculateCartTotal } from '../entities/cart';
 import { addCouponAtom, couponsAtom, deleteCouponAtom } from '../entities/coupon';
 import { addNotificationAtom } from '../entities/notification';
-import { calculateCartTotal } from '../models/cart';
 
 export function useCouponService() {
   const coupons = useAtomValue(couponsAtom);
