+++ src/advanced/pages/cart/ui/CartPage.tsx
@@ -4,11 +4,11 @@ import { CartItemList } from '../../../components/ui/CartItemList';
 import { CouponSelector } from '../../../components/ui/CouponSelector';
 import { PaymentSummary } from '../../../components/ui/PaymentSummary';
 import { ProductList } from '../../../components/ui/ProductList';
+import { calculateCartTotal } from '../../../entities/cart';
 import { productsAtom } from '../../../entities/product';
 import { useCartService } from '../../../hooks/useCartService';
 import { useCouponService } from '../../../hooks/useCouponService';
 import { useDebouncedSearch } from '../../../hooks/useDebouncedSearch';
-import { calculateCartTotal } from '../../../models/cart';
 import { Icon } from '../../../shared/icon';
 import { CartHeader } from '../../../widgets/cart-header';
